requirements.txt
pytelegrambotapi
python-dotenv
sqlalchemy
